refactor of the code into directories
