+++ config/default.conf
@@ -2,8 +2,9 @@
 server {
     listen 8080;
     listen 127.0.0.1:8081;
-    listen 192.168.1.1; # Should default to port 8080
+    listen 127.0.0.1:8081;
 
+    listen 192.168.1.1; # Should default to port 8080
     server_name example.com www.example.com sub.example.com;
     
     root /var/www/html;
@@ -13,13 +14,15 @@ server {
     client_max_body_size 20M; # Uppercase M
 
     error_page 403 404 /errors/40x.html;
-    error_page 500 /errors/500.html;
+    error_page 500 /errors/40x.html;
     error_page 502 503 504 /errors/50x.html;
 }
 
 # Server 2: Minimalist server with different values
 server {
+    listen 8080;
     listen 9000;
+
     
     server_name test.local;
     
@@ -47,7 +50,7 @@ server {
 
 # Server 4: Intentionally broken for testing validation 
 server {
-    listen 80;
+    listen 8070;
 
     root /tmp/www;
     # root /tmp/other; # Should throw "Duplicate 'root' directive" error
