+++ src/parser/config/ServerConfig.cpp
@@ -1,5 +1,4 @@
-#include "../inc/ServerConfig.hpp"
-#include <cstddef>
+#include "../../../inc/parser/config/ServerConfig.hpp"
 
 ServerConfig::ServerConfig() : clientMaxBodySize(-1) {}
 
@@ -10,6 +9,8 @@ ServerConfig::ServerConfig(const ServerConfig &other) { *this = other; }
 ServerConfig &ServerConfig::operator=(const ServerConfig &other) {
   if (this != &other) {
     this->listen = other.listen;
+    this->ips = other.ips;
+    this->ports = other.ports;
     this->serverNames = other.serverNames;
     this->root = other.root;
     this->clientMaxBodySize = other.clientMaxBodySize;
@@ -19,9 +20,15 @@ ServerConfig &ServerConfig::operator=(const ServerConfig &other) {
   return *this;
 }
 
-const std::vector<std::string> &ServerConfig::getListen() const {
+const std::vector<std::string> &ServerConfig::getListens() const {
   return this->listen;
 }
+
+const std::vector<std::string> &ServerConfig::getIPs() const {
+  return this->ips;
+}
+
+const std::vector<int> &ServerConfig::getPorts() const { return this->ports; }
 
 const std::vector<std::string> &ServerConfig::getServerNames() const {
   return this->serverNames;
@@ -41,12 +48,20 @@ const std::map<int, std::string> &ServerConfig::getErrorPages() const {
   return this->errorPages;
 }
 
-void ServerConfig::setListen(const std::string &listen_directive) {
+void ServerConfig::addListen(const std::string &listen_directive) {
   this->listen.push_back(listen_directive);
 }
 
-void ServerConfig::setServerName(const std::string &serverName) {
+void ServerConfig::addServerName(const std::string &serverName) {
   this->serverNames.push_back(serverName);
+}
+
+void ServerConfig::addIP(const std::string &ip) { this->ips.push_back(ip); }
+
+void ServerConfig::addPort(const int &port) { this->ports.push_back(port); }
+
+void ServerConfig::addIndexFile(const std::string &file) {
+  this->indexFiles.push_back(file);
 }
 
 void ServerConfig::setRoot(const std::string &root) { this->root = root; }
@@ -55,10 +70,6 @@ void ServerConfig::setClientMaxBodySize(const long &size) {
   this->clientMaxBodySize = size;
 }
 
-void ServerConfig::setIndexFile(const std::string &file) {
-  this->indexFiles.push_back(file);
-}
-
-void ServerConfig::setErrorPage(int code, const std::string &path) {
+void ServerConfig::addErrorPage(int code, const std::string &path) {
   this->errorPages[code] = path;
 }
